+++ pkg/sass-parser/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## 0.2.6
 
-* No user-visible changes.
+* Add `BooleanExpression` and `NumberExpression`.
 
 ## 0.2.5